Rename and add documentation
