+++ src/utils/caching/hybridCache.ts
@@ -1,5 +1,6 @@
 import { createLogger } from '@unly/utils-simple-logger';
 import deepmerge from 'deepmerge';
+import getTimestampsElapsedTime from '../time/getTimestampsElapsedTime';
 import { CachedItem, HybridCacheStorage, StorageOptions } from './hybridCacheStorage';
 
 const fileLabel = 'utils/cache/cache';
@@ -7,20 +8,20 @@ const logger = createLogger({ // eslint-disable-line no-unused-vars,@typescript-
   label: fileLabel,
 });
 
-type Options = {
-  ttl?: number; // In seconds
+type HybridCacheOptions = {
+  ttl?: number; // Default: 30 (seconds). If 0 then cache won't be invalidated (no TTL)
   storage?: {
-    type: 'memory';
+    type: 'memory'; // Default
   } | {
     type: 'disk';
     options: {
       filename: string;
     };
   };
-  enabled: boolean;
+  enabled: boolean; // Enabled by default
 }
 
-const defaultOptions: Required<Options> = {
+const defaultHybridCacheOptions: Required<HybridCacheOptions> = {
   ttl: 30,
   storage: {
     type: 'memory',
@@ -29,28 +30,45 @@ const defaultOptions: Required<Options> = {
 };
 
 /**
- * Returns the elapsed time between two timestamps, in seconds
+ * Hybrid cache, can use in-memory storage or on-disk storage.
  *
- * @param oldTimestamp
- * @param newTimestamp
+ * You can select the kind of storage you want through the options.
+ *
+ * Algorithm:
+ *  - Check if cache exist for key, if not then call the dataResolver and cache the result.
+ *  - If cache exists and record age is lower than TTL (default: 30 seconds), then return cached data. Otherwise, call the dataResolver and cache the result.
+ *
+ * Special considerations:
+ *  - If options.ttl is 0, then TTL is disabled and data will live indefinitely
+ *  - Memory cache storage is selected by default
+ *  - Memory cache storage will not work if you want to cache data at build time, when building the pages (typically: SSG). You must use the disk for this kind of usage.
+ *  - Disk cache storage works on Vercel, you need a writable file system
+ *
+ * @example With default (in-memory, TTL 30 secs)
+ *  const stuff = await hybridCache('CustomerTable', async () => await fetchMyStuff());
+ *
+ * @example Using disk storage with no TTL, meant to cache stuff once for all pages during server initial build (for SSG, typically - IS_SERVER_INITIAL_BUILD comes from next.config.js)
+ *  const stuff = await hybridCache('CustomerTable', async () => await fetchMyStuff(), { enabled: !!process.env.IS_SERVER_INITIAL_BUILD && process.env.NODE_ENV !== 'development', storage: {type: 'disk', options: {filename: 'my-stuff'}} });
+ *
+ * @param keyResolver string | Function Key used to store the data in the cache
+ * @param dataResolver Function Has to be implemented by the caller and is supposed to take care of fetching the data
+ * @param options HybridCacheOptions
  */
-const getTimestampsElapsedTime = (oldTimestamp: number, newTimestamp: number): number => (newTimestamp - oldTimestamp) / 1000;
-
-const hybridCache = async <T>(keyResolver: string | (() => string), fct: () => T, options: Partial<Options> = defaultOptions): Promise<T> => {
-  const { ttl, enabled, storage } = deepmerge(defaultOptions, options);
+const hybridCache = async <T>(keyResolver: string | (() => string), dataResolver: () => T, options: Partial<HybridCacheOptions> = defaultHybridCacheOptions): Promise<T> => {
+  const { ttl, enabled, storage } = deepmerge(defaultHybridCacheOptions, options);
 
   if (!enabled) { // Bypasses cache completely
     // eslint-disable-next-line no-console
     console.debug('Cache is disabled, bypassing');
-    return fct();
+    return dataResolver();
   }
   let cacheStorage: HybridCacheStorage;
   let storageOptions: StorageOptions = {};
 
   if (storage.type === 'memory') {
-    cacheStorage = require('./inMemoryCacheStorage');
+    cacheStorage = require('./memoryCacheStorage');
   } else {
-    cacheStorage = require('./inDiskCacheStorage');
+    cacheStorage = require('./diskCacheStorage');
     const { options } = storage;
     storageOptions = options;
   }
@@ -68,7 +86,7 @@ const hybridCache = async <T>(keyResolver: string | (() => string), fct: () => T
   if (typeof cachedItem !== 'undefined') {
     const { value, timestamp } = cachedItem;
 
-    if (getTimestampsElapsedTime(timestamp, +new Date()) < ttl) {
+    if (timestamp === 0 || getTimestampsElapsedTime(timestamp, +new Date()) < ttl) {
       return value;
     } else {
       logger.debug('Cache key has expired');
@@ -77,7 +95,7 @@ const hybridCache = async <T>(keyResolver: string | (() => string), fct: () => T
     logger.debug('Cache key does not exist');
   }
 
-  const unMemoizedResult: T = await fct();
+  const unMemoizedResult: T = await dataResolver();
   return cacheStorage.set(key, unMemoizedResult, storageOptions);
 };
